# Improve Enumerable.Reverse() performance