--- a/src/System.Linq/src/System/Linq/IReverseProvider.cs
+++ b/src/System.Linq/src/System/Linq/IReverseProvider.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+namespace System.Linq
+{
+    internal interface IReverseProvider<TElement> : IEnumerable<TElement>
+    {
+        IEnumerable<TElement> Reverse();
+    }
+}