… src/System.Linq/src/System/Linq/Select.cs
… src/System.Linq/src/System/Linq/Where.cs