- In sequence reverse, if the sequence is IList, we can enumerate the sequence backwards to improve performance and reduce memory allocation.

- In IEnumerable.Reverse().Reverse() case, we can return the source sequence direct to improve performance.

Here is the performance test case.
``` C#
    [MemoryDiagnoser]
    public class ReversePerfTest
    {
        private int[] data;

        [Params(0, 10, 100, 1000)]
        public int COUNT;


        [GlobalSetup]
        public void Setup()
        {
            data = new int[COUNT];
            for (int i = 0; i < COUNT; i++)
            {
                data[i] = i;
            }
        }

        [Benchmark]
        public int[] ReversePerf()
        {
            return data.ReversePerf().Skip(5).Take(10).ToArray();
        }

        [Benchmark(Baseline = true)]
        public int[] Reverse()
        {
            return data.Reverse().Skip(5).Take(10).ToArray();
        }
    }
```

Here is the result.
``` ini

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.765 (1803/April2018Update/Redstone4)
Intel Core i7-7500U CPU 2.70GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
Frequency=2835938 Hz, Resolution=352.6170 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0


```
|      Method | COUNT |      Mean |      Error |     StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------ |----------:|-----------:|-----------:|------:|-------:|------:|------:|----------:|
| **ReversePerf** |     **0** |  **68.00 ns** |  **0.9577 ns** |  **0.8489 ns** |  **0.45** | **0.0151** |     **-** |     **-** |      **32 B** |
|     Reverse |     0 | 150.68 ns |  2.9447 ns |  2.7544 ns |  1.00 | 0.0474 |     - |     - |     100 B |
|             |       |           |            |            |       |        |       |       |           |
| **ReversePerf** |    **10** | **107.81 ns** |  **1.8778 ns** |  **1.5680 ns** |  **0.28** | **0.0457** |     **-** |     **-** |      **96 B** |
|     Reverse |    10 | 390.91 ns |  7.5755 ns |  7.0861 ns |  1.00 | 0.1216 |     - |     - |     256 B |
|             |       |           |            |            |       |        |       |       |           |
| **ReversePerf** |   **100** | **130.14 ns** |  **2.9506 ns** |  **4.4163 ns** |  **0.26** | **0.0703** |     **-** |     **-** |     **148 B** |
|     Reverse |   100 | 508.18 ns | 10.0186 ns | 13.7136 ns |  1.00 | 0.3233 |     - |     - |     680 B |
|             |       |           |            |            |       |        |       |       |           |
| **ReversePerf** |  **1000** | **131.47 ns** |  **2.5980 ns** |  **2.5516 ns** |  **0.17** | **0.0703** |     **-** |     **-** |     **148 B** |
|     Reverse |  1000 | 768.03 ns | 15.0081 ns | 20.5433 ns |  1.00 | 2.0399 |     - |     - |    4280 B |