@stephentoub 2019-07-10
Thanks.

Can you please share real-world examples where this makes a meaningful difference?  There are a huge number of possible optimizations that can be made in LINQ, but most of them (like this one) requires a non-trivial amount of additional code that has the potential to be buggy, needs to be maintained (and updated when further additional interacting optimizations come along), adds potentially significant runtime overheads in terms of additional generic instantiations, etc... so we're only proceeding to add such optimizations when they're really worth they're weight and when it can be demonstrated that they make a meaningful difference on real-world cases.

Also, have you tried this on collections where indexing isn't O(1)? e.g. have you tried your perf test with an `ImmutableList<T>`?