@timandy 2019-07-11
@stephentoub  I test the ImmutableList<T> by this code.
```C#
    [MemoryDiagnoser]
    public class ReverseImutableList
    {
        private IEnumerable<int> data;

        [Params(0, 10, 100, 1000)]
        public int COUNT;


        [GlobalSetup]
        public void Setup()
        {
            int[] tmp = new int[COUNT];
            for (int i = 0; i < COUNT; i++)
            {
                tmp[i] = i;
            }

            ImmutableList<int> list = ImmutableList<int>.Empty;
            data = list.AddRange(tmp);
        }

        [Benchmark]
        public int ReversePerf()
        {
            int sum = 0;
            foreach (int item in data.ReversePerf())
                sum += item;
            return sum;
        }

        [Benchmark(Baseline = true)]
        public int Reverse()
        {
            int sum = 0;
            foreach (int item in data.Reverse())
                sum += item;
            return sum;
        }
    }

```

And here is the test result
``` ini

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.860 (1803/April2018Update/Redstone4)
Intel Core i7-7500U CPU 2.70GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
Frequency=2835940 Hz, Resolution=352.6168 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0


```
|      Method | COUNT |         Mean |        Error |       StdDev |       Median | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------ |-------------:|-------------:|-------------:|-------------:|------:|--------:|-------:|------:|------:|----------:|
| **ReversePerf** |     **0** |     **46.25 ns** |     **1.658 ns** |     **4.705 ns** |     **45.75 ns** |  **0.66** |    **0.08** | **0.0152** |     **-** |     **-** |      **32 B** |
|     Reverse |     0 |     70.85 ns |     2.420 ns |     6.745 ns |     70.06 ns |  1.00 |    0.00 | 0.0132 |     - |     - |      28 B |
|             |       |              |              |              |              |       |         |        |       |       |           |
| **ReversePerf** |    **10** |    **248.58 ns** |     **7.148 ns** |    **20.623 ns** |    **242.80 ns** |  **0.37** |    **0.05** | **0.0150** |     **-** |     **-** |      **32 B** |
|     Reverse |    10 |    673.63 ns |    21.889 ns |    64.198 ns |    659.91 ns |  1.00 |    0.00 | 0.0372 |     - |     - |      80 B |
|             |       |              |              |              |              |       |         |        |       |       |           |
| **ReversePerf** |   **100** |  **2,960.12 ns** |    **95.955 ns** |   **275.312 ns** |  **2,889.61 ns** |  **0.80** |    **0.08** | **0.0114** |     **-** |     **-** |      **32 B** |
|     Reverse |   100 |  3,714.60 ns |    73.544 ns |   105.475 ns |  3,685.10 ns |  1.00 |    0.00 | 0.2060 |     - |     - |     440 B |
|             |       |              |              |              |              |       |         |        |       |       |           |
| **ReversePerf** |  **1000** | **50,477.70 ns** |   **969.233 ns** | **1,077.300 ns** | **50,799.48 ns** |  **1.36** |    **0.08** |      **-** |     **-** |     **-** |      **32 B** |
|     Reverse |  1000 | 37,790.31 ns | 1,011.487 ns | 2,852.917 ns | 36,981.29 ns |  1.00 |    0.00 | 1.8921 |     - |     - |    4041 B |



ps: ListPartition also access IList by index, should be optimize?