@adamsitnik 2019-07-31
Hi @timandy 

Thank you for your PR!

> IEnumerable.Reverse().Reverse()

This is a performance anti-pattern, we should not be optimizing for such scenario

> DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0

I can see that you've run the benchmarks for Full .NET Framework 4.7.2 **32bit**. Could you please run the benchmarks using a local build of CoreFX? [This](https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-corefx.md#corefx-prerequisites) doc describes how to run the benchmarks for private build of CoreFX.

Once you build CoreFX in Release and have a copy of CoreRun stored in `$pathToCoreRun`

```cs
class Program
{
    static void Main(string[] args) 
          => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
}
```

```cmd
dotnet run -c Release -f netcoreapp3.0 --filter * --corerun $pathToCoreRun
```